% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_scores.R
\name{activity_scores}
\alias{activity_scores}
\title{Activity Scores}
\usage{
activity_scores(
  obj,
  k = 1,
  mcmc.use = NULL,
  norm = FALSE,
  plot_it = FALSE,
  ...
)
}
\arguments{
\item{obj}{Either a BASS model, a matrix, or a list of matrices (with one element per MCMC iteration)}

\item{k}{The number of columns of W to consider}

\item{mcmc.use}{a vector of indices telling which MCMC draws to use}

\item{norm}{Logical, should activity scores be normalized to have length one?}

\item{plot_it}{Logical, should a plot be made?}

\item{...}{additional arguments passed to the corresponding `C_bass` function.}
}
\value{
the activity scores
}
\description{
This function computes the activity scores for main effects of the variables
}
