% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bassfunc2bass.R
\name{bassfunc2bass}
\alias{bassfunc2bass}
\title{Convert functional BASS model to BASS model}
\usage{
bassfunc2bass(bfm)
}
\arguments{
\item{bfm}{an object of class bass, where a functional variable has been specified.}
}
\description{
The argument to this function is the output of a \code{bass()} call when a single functional variable is specified using the \code{xx.func} argument.
Note that the resulting model may not be a valid bass object for some applications,
but the resulting model can be passed to \code{concordance::C_bass()} and related functions.
}
\examples{
# simulate data (Friedman function with first variable as functional)
f<-function(x){
10*sin(pi*x[,1]*x[,2])+20*(x[,3]-.5)^2+10*x[,4]+5*x[,5]
}
sigma <- 1 # noise sd
n <- 500 # number of observations
nfunc <- 50 # size of functional variable grid
xfunc <- seq(0,1,length.out=nfunc) # functional grid
x <- matrix(runif(n*9),n,9) # 9 non-functional variables, only first 4 matter
X <- cbind(rep(xfunc,each=n),kronecker(rep(1,nfunc),x)) # to get y
y <- matrix(f(X),nrow=n)+rnorm(n*nfunc,0,sigma)

# fit BASS
mod_func <- bass(x,y,xx.func=xfunc)

# convert to standard BASS model
mod <- bassfunc2bass(mod_func)

# Estimate C matrix (augmentation approach)
C <- C_bass(mod)
}
