% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bassfunc2bass_fixed_t.R
\name{bassfunc2bass_fixed_t}
\alias{bassfunc2bass_fixed_t}
\title{Extract scalar BASS model from a functional BASS model for fixed functional variable values}
\usage{
bassfunc2bass_fixed_t(bassfunc, func.use, verbose = FALSE)
}
\arguments{
\item{bassfunc}{A BASS model object (with functional variables).}

\item{func.use}{A numeric vector of fixed values for the functional variable \( t \).}

\item{verbose}{logical; should progress be displayed?}
}
\value{
A list of BASS models for each \( t \)-value (or a single model if length(func.use) == 1).
}
\description{
Extracts the BASS model(s) corresponding to fixed values of the functional variable.
For each specified \( t \)-value, calculates the contribution of the functional variable
to the tensor product basis functions and returns a modified BASS model.
}
\examples{
# Simulate bass models for each of the three cases
f<-function(x){
  10*sin(pi*x[,1]*x[,2])+20*(x[,3]-.5)^2+10*x[,4]+5*x[,5]
}
n<-500 # number of observations
nfunc<-50 # size of functional variable grid
xfunc<-seq(0,1,length.out=nfunc) # functional grid
x<-matrix(runif(n*9),n,9) # 9 non-functional variables, only first 4 matter
X<-cbind(rep(xfunc,each=n),kronecker(rep(1,nfunc),x)) # to get y
y<-matrix(f(X),nrow=n)

# Scalar response
X <- lhs::randomLHS(500, 5)
y <- f(X)
bassfunc <- bass(x, y, xx.func=xfunc)

# Extract the BASS model for t = 0.5
mod_fixed <- funcbass2bass_fixed_t(mod_func, func.use = 0.5)

# Extract models for multiple t values
mods_fixed <- funcbass2bass_fixed_t(mod_func, func.use = c(0.5, 0.75))
}
