% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcbass2bass.R
\name{basslc2bass}
\alias{basslc2bass}
\title{Convert a linear combination of BASS models to a single BASS model}
\usage{
basslc2bass(
  mod_list,
  weights = rep(1, length(mod_list)),
  offset = 0,
  yy = NULL,
  mcmc.use = NULL
)
}
\arguments{
\item{mod_list}{A list of bass models.}

\item{weights}{An optional vector of weights.}

\item{offset}{This value is added to the intercept of the resulting model.}

\item{yy}{The data vector. When unspecified, y is taken as the linear combination of responses for the individual bass models.}

\item{mcmc.use}{set of indices telling which mcmc draws to use.}
}
\description{
A linear combination of BASS models is also a BASS model. This function takes a list of BASS models (all with the same data matrix xx.des)
and returns the resulting linear combination as a new BASS model.
One useful application of this function is to convert bassPCA to bass for a fixed time point.
Does not currently work for bass models with functional or categorical inputs.
}
\examples{
#' # simulate data (Friedman function with first variable as functional)
n <- 500
p <- 2
X <- matrix(runif(n*p), ncol=p)
y1 <- apply(X, 1, duqling::dms_additive)
y2 <- apply(X, 1, duqling::dms_harmonic)
z <- y1 + y2

# Fit bass models
mod1 <- bass(X, y1)
mod2 <- bass(X, y2)
mod_list <- list(mod1, mod2)

# Combine models
mod <- basslc2bass(mod_list, yy=z)

# Get predictions
yhat1 <- colMeans(predict(mod1, X))
yhat2 <- colMeans(predict(mod2, X))
yhat <- colMeans(predict(mod, X))

plot(yhat1 + yhat2, yhat)

}
