% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bassPCA2bass_fixed_t.R
\name{bassPCA2bass_fixed_t}
\alias{bassPCA2bass_fixed_t}
\title{Extract scalar BASS model from bassPCA model for fixed functional variable values}
\usage{
bassPCA2bass_fixed_t(bassPCA, func.use, func.true = NULL)
}
\arguments{
\item{bassPCA}{A bassBasis model object (from bassPCA() function).}

\item{func.use}{A numeric vector of fixed values for the functional variable \( t \).}

\item{func.true}{An optional vector of values for the functional variable in bassPCA. Should have length equal to nrow(bassPCA$dat$basis).}
}
\value{
A list of BASS models for each \( t \)-value (or a single model if length(func.use) == 1).
}
\description{
Extracts the BASS model(s) corresponding to fixed values of the functional variable.
For each specified \( t \)-value, calculates the contribution of the functional variable
to the tensor product basis functions and returns a modified BASS model.
}
\details{
Since bassPCA doesn't
}
\examples{
f<-function(x, t){
10*sin(pi*t*x[1])+20*(x[2]-.5)^2+10*x[3]+5*x[4]
}
XX <- lhs::randomLHS(500, 5)
y1 <- apply(XX, 1, f, t=0.5)
xfunc <- seq(0, 1, length.out=20)
yfunc <- t(apply(XX, 1, f, t=xfunc))

# Fit a bassPCA model
mod3_full <- bassPCA(XX, yfunc)

# Extract univariate model at t = 0.5
mod3 <- bassPCA2bass_fixed_t(mod3_full, 0.5)
}
