% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_scores.R
\name{coactivity_scores}
\alias{coactivity_scores}
\title{Co-Activity Scores}
\usage{
coactivity_scores(
  obj1,
  obj2 = NULL,
  k = 1,
  signed = TRUE,
  mcmc.use = NULL,
  norm = FALSE,
  plot_it = TRUE,
  ...
)
}
\arguments{
\item{obj1}{Either a BASS model, a matrix, or a list of matrices (with one element per MCMC iteration).}

\item{obj2}{Either a BASS model or NULL. If NULL, then \code{obj1} should be a matrix (or list of matrices) corresponding to the Vfg matrix.}

\item{k}{The number of columns of W to consider}

\item{signed}{Use signed or unsigned version of coactivity scores?}

\item{mcmc.use}{a vector of indices telling which MCMC draws to use}

\item{norm}{Logical, should activity scores be normalized to have length one?}

\item{plot_it}{Logical, should a plot be made?}

\item{...}{additional arguments passed to the corresponding `Cfg_bass` function.}
}
\value{
the co-activity scores
}
\description{
This function computes the co-activity scores for main effects of the variables
}
