% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_scores.R
\name{concordance}
\alias{concordance}
\title{Concordance}
\usage{
concordance(obj1, obj2, Cfg = NULL, mcmc.use = NULL, ...)
}
\arguments{
\item{obj1}{Either a BASS model or a matrix (or a list of matrices) with one element per MCMC iteration, corresponding to Cf.}

\item{obj2}{Either a BASS model or a matrix (or a list of matrices) with one element per MCMC iteration, corresponding to Cg.}

\item{Cfg}{A matrix or list of matrices corresponding to Cfg. Required if obj1/obj2 are not both BASS models.}

\item{mcmc.use}{a vector of indices telling which MCMC draws to use}

\item{...}{additional arguments passed to the corresponding `C_bass` or `Cfg_bass` function.}
}
\value{
A scalar or vector of concordance scores
}
\description{
This function computes the concordance between two models
}
