% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_bass.R
\name{E_bass}
\alias{E_bass}
\title{Expected value of a BASS model}
\usage{
E_bass(mod, prior = NULL, mcmc.use = NULL, func.use = NULL, verbose = FALSE)
}
\arguments{
\item{mod}{a fitted BASS model. The output of the bass() or bassPCA() functions.}

\item{prior}{a list, like one returned by the \code{build_prior()} function. See the documentation for details.}

\item{mcmc.use}{a vector of indices telling which mcmc draws to use}

\item{func.use}{a vector indicating which values of the functional variable to compute C for, if applicable}

\item{verbose}{Doesn't do anything currently.}
}
\value{
A list representing the posterior distribution of the Constantine matrix.
}
\description{
Closed form estimator of the expected value of a BASS function
}
\details{
Returns the expected value of a BASS model.
}
\examples{
# FRIEDMAN FUNCTION
# First input is treated as functional
# Use p=5, so there is one inert variable
f <- function(x, t) {
  10 * sin(pi * t * x[1]) + 20 * (x[2] - 0.5)^2 + 10 * x[3] + 5 * x[4]
}

# ===========================================
#        GENERATE DATA
# ===========================================
XX <- lhs::randomLHS(500, 5)
y1 <- apply(XX, 1, f, t = 0.5)
mod <- bass(XX, y1)
E <- E_bass(mod)
}
