% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_bass.R
\name{gradient_bass}
\alias{gradient_bass}
\title{Gradient evaluations of a BASS model}
\usage{
gradient_bass(object, newdata, mcmc.use = NULL, verbose = FALSE)
}
\arguments{
\item{object}{A fitted model, output from the bass function}

\item{newdata}{A matrix of input locations at which to evaluate the gradient. The columns should correspond to the same variables used in the \code{bass} function.}

\item{mcmc.use}{A vector indexing which MCMC iterations to be used.}

\item{verbose}{logical; should progress be displayed}
}
\value{
An array of gradient evaluations. First dimension indexes MCMC iteration, second dimension indexes input location, third dimension indexes the input variable.
}
\description{
This function behaves similarly to the predict.bass method, except that it returns
predictions for the gradient of f.
}
\examples{
#' # simulate data (Friedman function with first variable as functional)
n <- 500
p <- 2
X <- matrix(runif(n*p), ncol=p)
y <- apply(X, 1, duqling::dms_additive)
mod <- bass(X, y)

Xnew <-  matrix(runif(100*p), ncol=p)
gradients <- gradient_bass(mod, Xnew)

}
