% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_scores.R
\name{shapley_grad}
\alias{shapley_grad}
\title{Gradient-Based Shapley Values}
\usage{
shapley_grad(obj, mcmc.use = NULL, norm = FALSE, plot_it = FALSE, ...)
}
\arguments{
\item{obj}{Either a BASS model, a matrix, or a list of matrices (with one element per MCMC iteration)}

\item{mcmc.use}{a vector of indices telling which MCMC draws to use}

\item{norm}{Logical, should Shapley values be normalized to have length one?}

\item{plot_it}{Logical, should a plot be made?}

\item{...}{additional arguments passed to the corresponding `C_bass` function.}
}
\value{
the Shapley values
}
\description{
This function computes the Shapley Values (gradient-based) as described by Duan and Okten (2023).
}
\references{
Duan, Hui, and Giray Okten. "Derivative-based Shapley value for global sensitivity analysis and machine learning explainability." International Journal for Uncertainty Quantification 15.1 (2025).
}
