% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_bass.R
\name{Z_bass}
\alias{Z_bass}
\title{Expected gradient with BASS}
\usage{
Z_bass(
  mod,
  prior = NULL,
  mcmc.use = NULL,
  use_native_scale = FALSE,
  func.use = NULL,
  verbose = FALSE
)
}
\arguments{
\item{mod}{a fitted BASS model. The output of the bass() or bassPCA() functions.}

\item{prior}{a list, like one returned by the \code{build_prior()} function. See the documentation for details.}

\item{mcmc.use}{a vector of indices telling which mcmc draws to use}

\item{use_native_scale}{logical (default `TRUE`). Determines the scale of the inputs for computing the \( C \)-matrix. When `TRUE`, the \( C \)-matrix is computed on the original (native) scale of the input variables. When `FALSE`, the \( C \)-matrix corresponds to the inputs normalized to the \([0, 1]\) range, as used internally by BASS. This also affects derived quantities, such as activity scores..}

\item{func.use}{a vector indicating which values of the functional variable to compute C for, if applicable}

\item{verbose}{Doesn't do anything currently.}
}
\value{
A list representing the posterior distribution of the Constantine matrix.
}
\description{
Closed form estimator of the expected value of the gradient of a function
}
\details{
Returns the expected value of the gradient of a BASS model.The \code{use_native_scale} flag indicates whether the C matrix should be transformed to the native space before returning.
}
\examples{
# FRIEDMAN FUNCTION
# First input is treated as functional
# Use p=5, so there is one inert variable
f <- function(x, t) {
  10 * sin(pi * t * x[1]) + 20 * (x[2] - 0.5)^2 + 10 * x[3] + 5 * x[4]
}

# ===========================================
#        GENERATE DATA
# ===========================================
XX <- lhs::randomLHS(500, 5)
y1 <- apply(XX, 1, f, t = 0.5)
mod <- bass(XX, y1)
Z <- Z_bass(mod)
}
